Small refactoring of menus, mostly pushing existing code out into modules for easier refactoring later
