+++ src/Commands.js
@@ -20,5 +20,6 @@ define(function (require, exports, module) {
     exports.FILE_ADD_TO_WORKING_SET = "file.addToWorkingSet";
     exports.FILE_RELOAD = "file.reload";
     exports.FILE_QUIT = "file.quit";
+    exports.DEBUG_JSLINT = "debug.jslint";
 });
