+++ src/brackets.js
@@ -3,7 +3,7 @@
  */
 
 /*jslint vars: true, plusplus: true, devel: true, browser: true, nomen: true, indent: 4, maxerr: 50 */
-/*global define: false, brackets: true, $: false, JSLINT: false, PathUtils: false */
+/*global define: false, brackets: true, $: false, PathUtils: false */
 
 /**
  * brackets is the root of the Brackets codebase. This file pulls in all other modules as
@@ -22,13 +22,13 @@ define(function (require, exports, module) {
     require("widgets/bootstrap-dropdown");
     require("widgets/bootstrap-modal");
     require("thirdparty/path-utils/path-utils.min");
-    require("thirdparty/jslint/jslint");
     
     // Load dependent modules
     var ProjectManager          = require("ProjectManager"),
         DocumentManager         = require("DocumentManager"),
         EditorManager           = require("EditorManager"),
         WorkingSetView          = require("WorkingSetView"),
+        DebugCommandHandlers    = require("DebugCommandHandlers"),
         FileCommandHandlers     = require("FileCommandHandlers"),
         FileViewController      = require("FileViewController"),
         FileSyncManager         = require("FileSyncManager"),
@@ -37,7 +37,9 @@ define(function (require, exports, module) {
         Commands                = require("Commands"),
         CommandManager          = require("CommandManager"),
         CodeHintManager         = require("CodeHintManager"),
-        PerfUtils               = require("PerfUtils");
+        PerfUtils               = require("PerfUtils"),
+        Menus                   = require("Menus"),
+        JSLint                  = require("JSLint");
 
     // Define core brackets namespace if it isn't already defined
     //
@@ -217,7 +219,6 @@ define(function (require, exports, module) {
 
     // Main Brackets initialization
     $(document).ready(function () {
-        var _enableJSLint = true;
         
         function initListeners() {
             // Prevent unhandled drag and drop of files into the browser from replacing 
@@ -259,177 +260,7 @@ define(function (require, exports, module) {
        
         }
         
-        function runJSLint() {
-            var currentDoc = DocumentManager.getCurrentDocument();
-            var ext = currentDoc ? PathUtils.filenameExtension(currentDoc.file.fullPath) : "";
-            var lintResults = $("#jslint-results");
-            var goldStar = $("#gold-star");
-            
-            if (_enableJSLint && /^(\.js|\.htm|\.html)$/i.test(ext)) {
-                var text = currentDoc.getText();
-                
-                // If a line contains only whitespace, remove the whitespace
-                // This should be doable with a regexp: text.replace(/\r[\x20|\t]+\r/g, "\r\r");,
-                // but that doesn't work.
-                var i, arr = text.split("\n");
-                for (i = 0; i < arr.length; i++) {
-                    if (!arr[i].match(/\S/)) {
-                        arr[i] = "";
-                    }
-                }
-                text = arr.join("\n");
-                
-                var result = JSLINT(text, null);
-                
-                if (!result) {
-                    var errorTable = $("<table class='zebra-striped condensed-table'>")
-                                       .append("<tbody>");
-                    var selectedRow;
-                    
-                    JSLINT.errors.forEach(function (item, i) {
-                        if (item) {
-                            var makeCell = function (content) {
-                                return $("<td/>").text(content);
-                            };
-                            
-                            // Add row to error table
-                            var row = $("<tr/>")
-                                .append(makeCell(item.line))
-                                .append(makeCell(item.reason))
-                                .append(makeCell(item.evidence || ""))
-                                .appendTo(errorTable);
-                            
-                            row.click(function () {
-                                if (selectedRow) {
-                                    selectedRow.removeClass("selected");
-                                }
-                                row.addClass("selected");
-                                selectedRow = row;
-                                currentDoc.setCursor(item.line - 1, item.character - 1);
-                                EditorManager.focusEditor();
-                            });
-                        }
-                    });
-
-                    $("#jslint-results .table-container")
-                        .empty()
-                        .append(errorTable);
-                    lintResults.show();
-                    goldStar.hide();
-                } else {
-                    lintResults.hide();
-                    goldStar.show();
-                }
-            } else {
-                // JSLint is disabled or does not apply to the current file, hide
-                // both the results and the gold star
-                lintResults.hide();
-                goldStar.hide();
-            }
-            
-            EditorManager.resizeEditor();
-        }
         
-        function initMenus() {
-            // Implements the File menu items
-            $("#menu-file-new").click(function () {
-                CommandManager.execute(Commands.FILE_NEW);
-            });
-            $("#menu-file-open").click(function () {
-                CommandManager.execute(Commands.FILE_OPEN);
-            });
-            $("#menu-file-close").click(function () {
-                CommandManager.execute(Commands.FILE_CLOSE);
-            });
-            $("#menu-file-save").click(function () {
-                CommandManager.execute(Commands.FILE_SAVE);
-            });
-            $("#menu-file-quit").click(function () {
-                CommandManager.execute(Commands.FILE_QUIT);
-            });
-
-            // Implements the 'Run Tests' menu to bring up the Jasmine unit test window
-            var testWindow = null;
-            $("#menu-debug-runtests").click(function () {
-                if (testWindow) {
-                    try {
-                        testWindow.location.reload();
-                    } catch (e) {
-                        testWindow = null;  // the window was probably closed
-                    }
-                }
-
-                if (!testWindow) {
-                    testWindow = window.open("../test/SpecRunner.html");
-                    testWindow.location.reload(); // if it was opened before, we need to reload because it will be cached
-                }
-            });
-            
-            // Other debug menu items
-//            $("#menu-debug-wordwrap").click(function() {
-//                editor.setOption("lineWrapping", !(editor.getOption("lineWrapping")));
-//            });     
-            
-            $("#menu-debug-jslint").click(function () {
-                _enableJSLint = !_enableJSLint;
-                runJSLint();
-                $("#jslint-enabled-checkbox").css("display", _enableJSLint ? "" : "none");
-            });
-            
-            $("#menu-debug-show-perf").click(function () {
-                var perfHeader = $("<div class='modal-header' />")
-                    .append("<a href='#' class='close'>&times;</a>")
-                    .append("<h3 class='dialog-title'>Performance Data</h3>");
-                
-                var perfBody = $("<div class='modal-body' style='padding: 0' />");
-
-                var data = $("<table class='zebra-striped condensed-table' style='max-height: 600px; overflow: auto;'>")
-                    .append("<thead><th>Operation</th><th>Time (ms)</th></thead>")
-                    .append("<tbody />")
-                    .appendTo(perfBody);
-                
-                var makeCell = function (content) {
-                    return $("<td/>").text(content);
-                };
-                
-                var getValue = function (entry) {
-                    // entry is either an Array or a number
-                    // If it is an Array, return the average value
-                    if (Array.isArray(entry)) {
-                        var i, sum = 0;
-                        
-                        for (i = 0; i < entry.length; i++) {
-                            sum += entry[i];
-                        }
-                        return String(Math.floor(sum / entry.length)) + " (avg)";
-                    } else {
-                        return entry;
-                    }
-                };
-                    
-                var testName;
-                var perfData = PerfUtils.perfData;
-                for (testName in perfData) {
-                    if (perfData.hasOwnProperty(testName)) {
-                        // Add row to error table
-                        var row = $("<tr/>")
-                            .append(makeCell(testName))
-                            .append(makeCell(getValue(perfData[testName])))
-                            .appendTo(data);
-                    }
-                }
-                                                             
-                var perfDlog = $("<div class='modal hide' />")
-                    .append(perfHeader)
-                    .append(perfBody)
-                    .appendTo(document.body)
-                    .modal({
-                        backdrop: "static",
-                        show: true
-                    });
-            });
-        }
-
         function initCommandHandlers() {
             FileCommandHandlers.init($("#main-toolbar .title"));
         }
@@ -476,20 +307,10 @@ define(function (require, exports, module) {
     
         initListeners();
         initProject();
-        initMenus();
+        Menus.init();
         initCommandHandlers();
         initKeyBindings();
         initWindowListeners();
-        
-        $(DocumentManager).on("currentDocumentChange", function () {
-            runJSLint();
-        });
-        
-        $(DocumentManager).on("documentSaved", function (event, document) {
-            if (document === DocumentManager.getCurrentDocument()) {
-                runJSLint();
-            }
-        });
         
         PerfUtils.addMeasurement("Application Startup");
     });
